Add a Core-App-Extension-Safe subspec
The code in UIApplication+Extensions.swift (with implicit references
to Application.shared) is not safe for use in a Today Widget.
Including it in a project with widgets generates...

EITHER the following compile-time warning:

ld: warning: linking against a dylib which is not safe for use in
application extensions: <path follows>

OR a compile time-error that the methods are marked unavailable,
depending on how Cocoapods generates the project files, especially
with regard to the "Require Only App-Extension-Safe API" setting.

A project has been creating a subspec in a separate branch, but as
it's of general utility, I'd like to move the subspec into the main
branch.